Note screen contents are one single lazy column! Logo in top appbar is from mipmap! A bunch of Task refactoring!

+++ app/src/main/java/us/huseli/retain/data/NoteRepository.kt
@@ -1,26 +1,34 @@
 package us.huseli.retain.data
 
 import android.content.Context
-import android.graphics.ImageDecoder
 import android.net.Uri
-import android.os.Build
-import android.provider.MediaStore
-import androidx.compose.ui.graphics.asImageBitmap
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import us.huseli.retain.Constants.IMAGE_SUBDIR
 import us.huseli.retain.Enums.NoteType
 import us.huseli.retain.LogInterface
 import us.huseli.retain.Logger
+import us.huseli.retain.data.entities.BitmapImage
 import us.huseli.retain.data.entities.ChecklistItem
 import us.huseli.retain.data.entities.Image
-import us.huseli.retain.data.entities.ImageWithBitmap
 import us.huseli.retain.data.entities.Note
 import us.huseli.retain.data.entities.NoteCombined
+import us.huseli.retain.nextcloud.NextCloudEngine
+import us.huseli.retain.nextcloud.tasks.DownloadMissingImagesTask
+import us.huseli.retain.nextcloud.tasks.DownloadNoteImagesTask
+import us.huseli.retain.nextcloud.tasks.DownstreamSyncTask
+import us.huseli.retain.nextcloud.tasks.RemoveImageTask
+import us.huseli.retain.nextcloud.tasks.RemoveImagesTask
+import us.huseli.retain.nextcloud.tasks.RemoveNotesTask
+import us.huseli.retain.nextcloud.tasks.RemoveOrphanImagesTask
+import us.huseli.retain.nextcloud.tasks.TestNextCloudTaskResult
+import us.huseli.retain.nextcloud.tasks.UploadImageTask
+import us.huseli.retain.nextcloud.tasks.UploadMissingImagesTask
+import us.huseli.retain.nextcloud.tasks.UploadNoteTask
+import us.huseli.retain.nextcloud.tasks.UpstreamSyncTask
 import java.io.File
 import java.util.UUID
 import javax.inject.Inject
@@ -32,7 +40,7 @@ class NoteRepository @Inject constructor(
     private val noteDao: NoteDao,
     private val checklistItemDao: ChecklistItemDao,
     private val imageDao: ImageDao,
-    private val nextCloud: NextCloud,
+    private val nextCloudEngine: NextCloudEngine,
     private val ioScope: CoroutineScope,
     override val logger: Logger,
     private val database: Database,
@@ -41,9 +49,39 @@ class NoteRepository @Inject constructor(
     val notes: Flow<List<Note>> = noteDao.flowList()
     val checklistItems: Flow<List<ChecklistItem>> = checklistItemDao.flowList()
     val nextCloudNeedsTesting = MutableStateFlow(true)
+    val bitmapImages = MutableStateFlow<List<BitmapImage>>(emptyList())
 
     init {
         syncNextCloud()
+
+        ioScope.launch {
+            imageDao.flowList().collect { images ->
+                val newBitmapImages = bitmapImages.value.toMutableList()
+                val currentImages = newBitmapImages.map { it.image }
+
+                images.forEach { image ->
+                    // Match by content (no point in updating if old and new
+                    // contents are identical):
+                    if (!currentImages.contains(image)) {
+                        image.toBitmapImage(context)?.let { bitmapImage ->
+                            newBitmapImages.removeIf { it.image.filename == image.filename }
+                            newBitmapImages.add(bitmapImage)
+                        }
+                    }
+                }
+                newBitmapImages.removeIf { !images.contains(it.image) }
+                bitmapImages.value = newBitmapImages
+            }
+        }
+    }
+
+    private fun addDownloadedImage(image: Image) {
+        image.toBitmapImage(context)?.let {
+            val newBitmapImages = bitmapImages.value.toMutableList()
+
+            newBitmapImages.add(it)
+            bitmapImages.value = newBitmapImages
+        }
     }
 
     suspend fun deleteChecklistItem(item: ChecklistItem) {
@@ -55,7 +93,7 @@ class NoteRepository @Inject constructor(
         imageDao.delete(image)
         noteDao.touch(image.noteId)
         File(imageDir, image.filename).delete()
-        nextCloud.RemoveImageTask(image).run()
+        RemoveImageTask(nextCloudEngine, image).run()
     }
 
     suspend fun deleteNotes(ids: Collection<UUID>) {
@@ -65,30 +103,8 @@ class NoteRepository @Inject constructor(
             File(imageDir, image.filename).delete()
         }
         noteDao.delete(ids)
-        nextCloud.RemoveNotesAndImagesTask(ids, images).run()
-    }
-
-    fun flowImagesWithBitmap(noteId: UUID? = null): Flow<List<ImageWithBitmap>> {
-        val flowList = if (noteId != null) imageDao.flowList(noteId) else imageDao.flowList()
-
-        return flowList.map { images ->
-            val imagesWithBitmap = mutableListOf<ImageWithBitmap>()
-
-            images.forEach { image ->
-                Uri.fromFile(File(imageDir, image.filename))?.let { uri ->
-                    val bitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-                        ImageDecoder.decodeBitmap(ImageDecoder.createSource(context.contentResolver, uri))
-                    } else {
-                        @Suppress("DEPRECATION")
-                        MediaStore.Images.Media.getBitmap(context.contentResolver, uri)
-                    }
-                    bitmap?.let {
-                        imagesWithBitmap.add(ImageWithBitmap(image, it.asImageBitmap()))
-                    }
-                }
-            }
-            imagesWithBitmap
-        }
+        RemoveNotesTask(nextCloudEngine, ids).run()
+        RemoveImagesTask(nextCloudEngine, images).run()
     }
 
     fun flowNote(id: UUID): Flow<Note?> = noteDao.flow(id)
@@ -103,17 +119,33 @@ class NoteRepository @Inject constructor(
         imageDao.insert(noteId, filename, mimeType, width, height, size)
         noteDao.touch(noteId)
         imageDao.get(filename)?.let { image ->
-            nextCloud.UploadImageTask(image).run()
+            UploadImageTask(nextCloudEngine, image).run { result ->
+                if (!result.success) logError("Failed to upload image to Nextcloud", result.error)
+            }
         }
     }
 
     suspend fun listChecklistItems(noteId: UUID): List<ChecklistItem> = checklistItemDao.listByNoteId(noteId)
 
     private fun syncNextCloud() {
-        nextCloud.DownstreamSyncTask().run { downTask ->
+        // First try to download any locally missing images, because this has
+        // highest prio:
+        ioScope.launch {
+            val images = imageDao.list()
+            val missingImages = images.filter { !File(imageDir, it.filename).exists() }
+
+            DownloadMissingImagesTask(nextCloudEngine, missingImages).run(
+                onEachCallback = { image, result -> if (result.success) addDownloadedImage(image) },
+                onReadyCallback = null,
+            )
+        }
+
+        DownstreamSyncTask(nextCloudEngine).run { downTaskResult ->
+            if (!downTaskResult.success) logError("Failed to download notes from Nextcloud", downTaskResult.error)
+
             // Remote files have been fetched and parsed; now update DB where needed.
             ioScope.launch {
-                val remoteNotesMap = downTask.remoteNotesCombined.associateBy { it.id }
+                val remoteNotesMap = downTaskResult.remoteNotesCombined.associateBy { it.id }
                 val notes = noteDao.list()
                 val checklistItems = checklistItemDao.list()
                 val images = imageDao.list()
@@ -146,7 +178,13 @@ class NoteRepository @Inject constructor(
                     noteDao.upsert(it)
                     checklistItemDao.replace(it.id, it.checklistItems)
                     imageDao.replace(it.id, it.images)
-                    nextCloud.DownloadNoteImagesTask(it).run()
+                    DownloadNoteImagesTask(nextCloudEngine, it).run(
+                        onEachCallback = { image, result ->
+                            if (result.success) addDownloadedImage(image)
+                            else logError("Failed to download image from Nextcloud", result.error)
+                        },
+                        onReadyCallback = null
+                    )
                 }
                 if (remoteUpdated.isNotEmpty()) {
                     log(
@@ -156,7 +194,14 @@ class NoteRepository @Inject constructor(
                 }
 
                 // Now upload all notes that are new or updated locally:
-                nextCloud.UpstreamSyncTask(localUpdated).run()
+                UpstreamSyncTask(nextCloudEngine, localUpdated).run { result ->
+                    if (!result.success) {
+                        if (result.unsuccessfulCount > 0)
+                            logError("Failed to upload ${result.unsuccessfulCount} notes to Nextcloud", result.error)
+                        else
+                            logError("Failed to upload notes to Nextcloud", result.error)
+                    }
+                }
 
                 syncImages()
             }
@@ -168,21 +213,18 @@ class NoteRepository @Inject constructor(
         val imageFilenames = images.map { it.filename }
 
         // Upload any images that are missing/wrong on remote:
-        nextCloud.UploadMissingImagesTask(images).run()
-        // Also download any locally missing images:
-        val missingImages = images.filter { !File(imageDir, it.filename).exists() }
-        nextCloud.DownloadMissingImagesTask(missingImages).run()
+        UploadMissingImagesTask(nextCloudEngine, images).run { result ->
+            if (!result.success) logError("Failed to upload image to Nextcloud", result.error)
+        }
         // Delete any orphan image files, both locally and on Nextcloud:
         imageDir.listFiles()?.forEach { file ->
-            if (!imageFilenames.contains(file.name)) {
-                file.delete()
-            }
+            if (!imageFilenames.contains(file.name)) file.delete()
         }
-        nextCloud.RemoveOrphanImagesTask(keep = imageFilenames).run()
+        RemoveOrphanImagesTask(nextCloudEngine, keep = imageFilenames).run()
     }
 
-    fun testNextcloud(uri: Uri, username: String, password: String, onResult: (NextCloudTestResult) -> Unit) {
-        nextCloud.testClient(uri, username, password) { result -> onResult(result) }
+    fun testNextcloud(uri: Uri, username: String, password: String, onResult: (TestNextCloudTaskResult) -> Unit) {
+        nextCloudEngine.testClient(uri, username, password) { result -> onResult(result) }
     }
 
     suspend fun updateChecklistItems(items: Collection<ChecklistItem>) = checklistItemDao.update(items)
@@ -190,14 +232,17 @@ class NoteRepository @Inject constructor(
     suspend fun upsertNote(id: UUID, type: NoteType, title: String, text: String, showChecked: Boolean, colorIdx: Int) {
         noteDao.upsert(id, type, title, text, showChecked, colorIdx)
         noteDao.get(id)?.let { note ->
-            nextCloud.UploadNoteTask(
+            UploadNoteTask(
+                nextCloudEngine,
                 noteCombined = NoteCombined(
                     note = note,
                     checklistItems = checklistItemDao.listByNoteId(id),
                     images = imageDao.list(id),
                     databaseVersion = database.openHelper.readableDatabase.version
                 )
-            ).run()
+            ).run { result ->
+                if (!result.success) logError("Failed to upload note to Nextcloud", result.error)
+            }
         }
     }
 }
